refactored PriceConverter, improvements & optimizations
